[ui] Use different views for public/private routes
Removes the sidebar from the main app and moves it
to a separate view. The sidebar is loaded only
on the routes saved under 'private' and requires
authentication.
A '/login' public route is also created.
